% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eGauge.R
\name{eGauge}
\alias{eGauge}
\title{Gauge Charts}
\usage{
eGauge(value, label = NULL, center = c(0.5,0.5),
               radius = 0.75, angle = c(-45, 225),
               range = c(0, 100),  percent = F,
               size = NULL, theme = "default",opt=list(),...)
}
\arguments{
\item{value}{value for Gauge Chart}

\item{center}{center of the plot,(width,height) 0.5 means 50\%}

\item{radius}{radius of the plot}

\item{angle}{start and end angle of the Gauge Chart}

\item{range}{max and min of the Gauge Chart}

\item{percent}{add "\%" after the value or not}

\item{theme}{theme of the Echarts}
}
\value{
an htmlwidgets object which can be rendered by Browser.
}
\description{
ECharts style Gauge charts.
}
\examples{
SuccessRate = 80
eGauge(SuccessRate)
eGauge(80,label = "Success Rate",
       percent=T,theme="macarons")
eGauge(80,label = "Success Rate",angle = c(45,-45),
         percent=T,theme="macarons")
}

